+++ src/test/java/nextstep/ladder/biz/LadderGeneratorTest.java
@@ -1,9 +1,9 @@
 package nextstep.ladder.biz;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class LadderGeneratorTest {
